refactor(frame): extract _attach/_detach, default correct_saturated to True
- Move EmptyBackend into _backends.py and use it as the initial backend
  state before _attach resolves the real one
- Extract _attach() to consolidate the path/backend wiring that was
  duplicated between __init__ and save_as
- Extract _detach() and use it in subtract(inplace=True) and
  _detached_copy() instead of open-coding _path = None
- Change the int64 intermediate in subtract_arrays to int32 (sufficient
  for uint16 differences and avoids unnecessary widening)
- Default correct_saturated to True in Frame.subtract; add two tests at
  the Frame level covering the blown-pixel interpolation and confirming
  the new default

+++ src/celestack/frame/_image_ops.py
@@ -161,7 +161,7 @@ def subtract_arrays(
         ValueError: If the dtype is unsupported.
     """
     if np.issubdtype(left.dtype, np.unsignedinteger):
-        result = left.astype(np.int64) - right.astype(np.int64)
+        result = left.astype(np.int32) - right.astype(np.int32)
         limits = np.iinfo(left.dtype)
         clipped = np.clip(result, limits.min, limits.max).astype(left.dtype)
         if correct_saturated:
